### 2018-06-10 #770 @whedon
[PRE REVIEW]: unyt: Handle, manipulate, and convert data with units in Python
**Submitting author:** @ngoldbaum (<a href="http://orcid.org/0000-0001-5557-267X">Nathan Goldbaum</a>)
**Repository:** <a href="https://github.com/yt-project/unyt" target ="_blank">https://github.com/yt-project/unyt</a>
**Version:** v1.0.3
