### 2018-06-10 #770 @ngoldbaum
Not sure why it couldn’t find the images, they’re in the paper directory: https://github.com/yt-project/unyt/tree/master/paper
