### 2018-06-18 #770 @ngoldbaum
How about Stuart Mumford (has experience with yt.units and astropy.units, is on the reviewer list for JOSS), Marten van Kerkwijk (astropy and numpy developer, helps maintain astropy's Quantity class) or Trevor Bekolay (reviewed unit packages in Python at SciPy 2013).
