### 2018-06-20 8f520a3
store SI-prefixability metadata in the unit registry
