+++ unyt/_unit_lookup_table.py
@@ -60,160 +60,164 @@ import numpy as np
 
 default_unit_symbol_lut = {
     # base
-    "m": (1.0, dimensions.length, 0.0, r"\rm{m}"),
-    "g":  (1.0e-3, dimensions.mass, 0.0, r"\rm{g}"),
-    "s":  (1.0, dimensions.time, 0.0, r"\rm{s}"),
-    "K":  (1.0, dimensions.temperature, 0.0, r"\rm{K}"),
-    "radian": (1.0, dimensions.angle, 0.0, r"\rm{radian}"),
-    "A": (1.0, dimensions.current_mks, 0.0, r"\rm{A}"),
-    "cd": (1.0, dimensions.luminous_intensity, 0.0, r"\rm{cd}"),
-    "mol": (1.0 / amu_grams, dimensions.dimensionless, 0.0, r"\rm{mol}"),
+    "m": (1.0, dimensions.length, 0.0, r"\rm{m}", True),
+    "g":  (1.0e-3, dimensions.mass, 0.0, r"\rm{g}", True),
+    "s":  (1.0, dimensions.time, 0.0, r"\rm{s}", True),
+    "K":  (1.0, dimensions.temperature, 0.0, r"\rm{K}", True),
+    "radian": (1.0, dimensions.angle, 0.0, r"\rm{radian}", True),
+    "A": (1.0, dimensions.current_mks, 0.0, r"\rm{A}", True),
+    "cd": (1.0, dimensions.luminous_intensity, 0.0, r"\rm{cd}", True),
+    "mol": (1.0 / amu_grams, dimensions.dimensionless, 0.0, r"\rm{mol}", True),
 
     # some cgs
-    "dyne": (1.0e-5, dimensions.force, 0.0, r"\rm{dyn}"),
-    "erg":  (1.0e-7, dimensions.energy, 0.0, r"\rm{erg}"),
-    "Ba": (0.1, dimensions.pressure, 0.0, r"\rm{Ba}"),
-    "esu":  (1.0e-3**1.5, dimensions.charge_cgs, 0.0, r"\rm{esu}"),
-    "gauss": (0.1**0.5, dimensions.magnetic_field_cgs, 0.0, r"\rm{G}"),
-    "degC": (1.0, dimensions.temperature, -273.15, r"^\circ\rm{C}"),
-    "statA": (1.0e-3**1.5, dimensions.current_cgs, 0.0, r"\rm{statA}"),
+    "dyne": (1.0e-5, dimensions.force, 0.0, r"\rm{dyn}", True),
+    "erg":  (1.0e-7, dimensions.energy, 0.0, r"\rm{erg}", True),
+    "Ba": (0.1, dimensions.pressure, 0.0, r"\rm{Ba}", True),
+    "esu":  (1.0e-3**1.5, dimensions.charge_cgs, 0.0, r"\rm{esu}", True),
+    "gauss": (0.1**0.5, dimensions.magnetic_field_cgs, 0.0, r"\rm{G}", True),
+    "degC": (1.0, dimensions.temperature, -273.15, r"^\circ\rm{C}", True),
+    "statA": (1.0e-3**1.5, dimensions.current_cgs, 0.0, r"\rm{statA}", True),
     "statV": (0.1*1.0e-3**0.5, dimensions.electric_potential_cgs,
-              0.0, r"\rm{statV}"),
-    "statohm": (100.0, dimensions.resistance_cgs, 0.0, r"\rm{statohm}"),
-    "Mx": (1.0e-3**1.5, dimensions.magnetic_flux_cgs, 0.0, r"\rm{Mx}"),
+              0.0, r"\rm{statV}", True),
+    "statohm": (100.0, dimensions.resistance_cgs, 0.0, r"\rm{statohm}", True),
[…]
