### 2018-06-28 c08a335
Use dimensions of units to do CGS <-> MKS conversions.
This fixes issues with converting code E&M magnetic units between unit systems
