### 2018-06-28 70f3e7d
Allow taking scalars to dimensionless array powers
