### 2018-06-29 93f6337
refactor unit system sanitization
