### 2018-07-01 #770 @mhvk
@kyleniemeyer - with @Cadair already reviewing, I feel adding me would not add much and perhaps make the review too "astropy biased".
