### 2018-07-03 8a37b5e
use pyenv to set up test environment
