+++ tox.ini
@@ -2,9 +2,6 @@
 envlist = docs,begin,py27-min,py27-mindeps,py{27,34,35,36},end
 
 [travis]
-os =
-    linux: docs,py{py,27,34,35,36}
-    osx: py{27,36}
 python =
     3.6: docs, py36
     3.5: py35
@@ -72,7 +69,9 @@ commands =
 
 [testenv:docs]
 changedir = docs
-deps = sphinx
+deps =
+  sphinx
+  numpy
 commands =
     python -m sphinx -M html "." "_build" -W
 
