+++ README.rst
@@ -24,6 +24,10 @@ unyt
         :target: https://codecov.io/gh/yt-project/unyt
         :alt: Test Coverage
 
+.. image:: http://joss.theoj.org/papers/dbc27acb614dd33eb02b029ef20e7fe7/status.svg
+        :target: http://joss.theoj.org/papers/dbc27acb614dd33eb02b029ef20e7fe7
+        :alt: Code Paper
+
 |
 
  .. image:: docs/_static/yt_logo_small.png
