### 2018-07-09 #809 @kyleniemeyer
@tbekolay thank you! The version mismatch is fine, we recognize the software will change after submission (but hopefully not substantially).

Yes, we do ask that you fill out the checklist—that, plus the resolution of any issues you may bring up, is how we determine that our acceptance criteria is met. There aren't any additional criteria, though any suggestions/comments on the paper itself (or any other part of the software) are welcome.
