### 2018-07-13 #809 @ygrange
Ok, I finally got around to have a look at this. I'll try to submit some comments half-way next week.
