### 2018-07-14 5a538d4
make UnitRegistry.__contains__ a bit smarter
