### 2018-07-14 3814fc9
remove yt-specific error message, we can handle that inside yt
