+++ unyt/unit_registry.py
@@ -309,14 +309,8 @@ def _lookup_unit_symbol(symbol_str, unit_symbol_lut):
             return ret
 
     # no dice
-    if symbol_str.startswith('code_'):
-        raise UnitParseError(
-            "Code units have not been defined. \n"
-            "Try creating the array or quantity using ds.arr or ds.quan "
-            "instead.")
-    else:
-        raise UnitParseError("Could not find unit symbol '%s' in the provided "
-                             "symbols." % symbol_str)
+    raise UnitParseError("Could not find unit symbol '%s' in the provided "
+                         "symbols." % symbol_str)
 
 
 #: The default unit registry
