### 2018-07-19 #809 @ygrange
The main comment I have is: Great paper, I really enjoyed reading it! You did a really good job at analysing the differences in performance and functionality between unyt and the two comparable packages. I will go through the checklist shortly, but let me post my comments on the contents first:

First a comment on the code base: Your code coverage is really great! There is no documentation for users to execute the unit tests after installation from source. It's probably one extra line to the readme, but would certainly be great to have.

On the text then:
page 7: “for each of the benchmarks below” -> not all benchmarks are below this text so it’s a bit confusing. Rephrasing to "in this section" or something similar fixes this.

Let me repeat that I really love the thorough profiling here. Very impressive and very nice to read!

I have a few comments
- Figure 2: I don’t really get why the overhead is negative for square root of the largest arrays. Probably the effect is similar to what you discuss for figure 3 later.
- discussion of figure 3 and 4:
  - Using the phrase “overhead of ~0.5” is a bit confusing because my brain parses that as a 150% effect i.s.o. 50% which is meant. Also I find it a bit hard to parse what is exactly meant here. What really helped me was to realise that the effect is only present in figure 3 (different units) and not 4 (equal units). I believe that explicitly mentioning this can make it a bit easier to grasp.
  - The different handling of divisions essentially means that in astropy and Pint 1*kg / (1*kg) = 1kg while in unyt the answer is 1 (dimensionless), right? It may be illustrative to put down an example like this in the text.
- In the description of ufunc you talk about figure 4 and 5. Should be 5 and 6. (a bit futher Figure 4 should be Figure 5. Also, I assume the abovementioned effect is also present in the ufunc versions.
