### 2018-07-19 #809 @ygrange
On the dependencies: to do the testing, one needs to install a few extra dependencies.
