### 2018-07-19 #809 @ygrange
Documentation: the docstrings look really good and complete. Even though they are quite descriptive by nature I would advice to add a desciption to the exception classes (i.e. if I use unyt  in my work, what case is covere by what exception?)
