### 2018-07-19 #809 @ygrange
I see the list of authors in the paper.md file, but there is no author listed in the PDF. Is this a technical problem? (question for the @kyleniemeyer I guess)
