### 2018-07-20 #809 @Cadair
This is a really nice paper, with a great comparison between the three packages. The quality of the code, docs and tests all looks fantastic as I would've expected :smile:. I hope that this package gains a lot of traction by becoming such a lightweight standalone package (something I have thought `astropy.units` should do for a long time). I look forward to using it in some of my own code in the future.

One small comment on the paper:

* In the section on "Unary arithmetic operations" there is a line that reads "the y-axis is plotted on a log scale", should this not be "x-axis"?
