### 2018-07-20 #809 @kyleniemeyer
@ygrange regarding your question about the authors, I do see the authors right below the title in the PDF—do you see something else?
