### 2018-07-23 8cfa0c9
Add numpy and sympy import checks to top-level __init__.py
Closes #32
