+++ tox.ini
@@ -23,6 +23,7 @@ deps =
     pytest-cov
     pytest-doctestplus
     flake8
+    setuptools
     py27: backports.functools_lru_cache
 commands =
     pytest --cov=unyt --cov-append --doctest-modules --doctest-plus --doctest-glob='*.rst' --basetemp={envtmpdir}
