### 2018-07-23 a50e0e8
add python3.7 builds
