+++ .travis.yml
@@ -32,7 +32,8 @@ before_install:
   - pyenv install -s 3.4.8
   - pyenv install -s 3.5.5
   - pyenv install -s 3.6.5
-  - pyenv local 2.7.15 3.4.8 3.5.5 3.6.5
+  - pyenv install -s 3.7.0
+  - pyenv local 2.7.15 3.4.8 3.5.5 3.6.5 3.7.0
   - |
       if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then
         brew install python
