+++ tox.ini
@@ -1,8 +1,9 @@
 [tox]
-envlist = docs,begin,py27-min,py27-mindeps,py{27,34,35,36},end
+envlist = docs,begin,py27-min,py27-mindeps,py{27,34,35,36,37},end
 
 [travis]
 python =
+    3.7: py37
     3.6: docs, py36
     3.5: py35
     3.4: py34
