### 2018-07-23 6abf6b1
switch to xenial image so python3.7 builds
