+++ .travis.yml
@@ -1,6 +1,7 @@
 # Config file for automatic testing at travis-ci.org
 
-sudo: false
+sudo: true
+dist: xenial
 cache:
   pip: true
   directories:
