### 2018-07-23 db8fd51
ignore coverage for error messages since these are hard to test
