### 2018-07-25 a167ed4
Add spatial frequency as a dimension
