### 2018-07-25 7d46825
Add support for spatial frequency to spectral equivalence
