+++ unyt/equivalencies.py
@@ -24,6 +24,7 @@ from unyt.dimensions import (
     density,
     number_density,
     flux,
+    spatial_frequency,
 )
 from unyt.exceptions import InvalidUnitEquivalence
 
@@ -182,9 +183,9 @@ class MassEnergyEquivalence(Equivalence):
 class SpectralEquivalence(Equivalence):
     """Equivalence between wavelength, frequency, and energy of a photon.
 
-    Given a photon with wavelength :math:`\\lambda`, frequency :math:`\\nu`
-    and Energy :math:`E`, these quantities are related by the following
-    forumlae:
+    Given a photon with wavelength :math:`\\lambda`, spatial frequency
+    :math:`\\bar\\nu`, frequency :math:`\\nu` and Energy :math:`E`,
+    these quantities are related by the following forumlae:
 
     .. math::
 
@@ -195,7 +196,7 @@ class SpectralEquivalence(Equivalence):
     Example
     ------
     >>> print(SpectralEquivalence())
-    spectral: length <-> frequency <-> energy
+    spectral: length <-> spatial_frequency <-> frequency <-> energy
     >>> from unyt import angstrom, km
     >>> (3*angstrom).to_equivalent('keV', 'spectral')
     unyt_quantity(4.13280644, 'keV')
@@ -203,7 +204,7 @@ class SpectralEquivalence(Equivalence):
     unyt_quantity(0.29979246, 'MHz')
     """
     type_name = "spectral"
-    _dims = (length, rate, energy,)
+    _dims = (length, rate, energy, spatial_frequency)
 
     def _convert(self, x, new_dims):
         from unyt import physical_constants as pc
@@ -212,19 +213,32 @@ class SpectralEquivalence(Equivalence):
                 return np.divide(pc.clight*pc.hmks, x, out=self._get_out(x))
             elif x.units.dimensions == rate:
                 return np.multiply(x, pc.hmks, out=self._get_out(x))
+            elif x.units.dimensions == spatial_frequency:
+                return np.divide(x, pc.clight*pc.hmks, out=self._get_out(x))
         elif new_dims == length:
             if x.units.dimensions == rate:
                 return np.divide(pc.clight, x, out=self._get_out(x))
             elif x.units.dimensions == energy:
                 return np.divide(pc.hmks*pc.clight, x, out=self._get_out(x))
+            elif x.units.dimensions == spatial_frequency:
+                return np.divide(1, x, out=self._get_out(x))
         elif new_dims == rate:
             if x.units.dimensions == length:
                 return np.divide(pc.clight, x, out=self._get_out(x))
             elif x.units.dimensions == energy:
                 return np.divide(x, pc.hmks, out=self._get_out(x))
+            elif x.units.dimensions == spatial_frequency:
+                return np.divide(x, pc.clight, out=self._get_out(x))
+        elif new_dims == spatial_frequency:
+            if x.units.dimensions == length:
+                return np.divide(1, x, out=self._get_out(x))
+            elif x.units.dimensions == energy:
+                return np.divide(x, pc.hmks*pc.clight, out=self._get_out(x))
+            elif x.units.dimensions == rate:
+                return np.divide(x, pc.clight, out=self._get_out(x))
 
     def __str__(self):
-        return "spectral: length <-> frequency <-> energy"
+        return "spectral: length <-> spatial_frequency <-> frequency <-> energy"
 
 
 class SoundSpeedEquivalence(Equivalence):
