### 2018-07-25 cdc3436
Add tests and fix conversions to match
