### 2018-07-25 e7383f0
Add spatial frequency to places where spectral equivalence is mentioned
