### 2018-07-25 d792b96
Fix doctest example
