### 2018-07-26 c2addb3
mention nubar in latex formula
