### 2018-07-30 4368e25
add docs for unyt.test
