### 2018-07-30 fb717b1
Avoid confusion about placement of benchmark figures
This is in response to the following comment:

> page 7: “for each of the benchmarks below” -> not all benchmarks are below this text so it’s a bit confusing. Rephrasing to "in this section" or something similar fixes this.

The fix is just to not say "below". This avoids confusion due to some of the
figures appearing before this statement.
