### 2018-07-30 4947f46
fix confusion about distinction between an overhead measurement equal to 1 and equal to zero
