### 2018-07-30 83e0200
fix minor wording issues
