### 2018-07-30 8cc11b5
Adjust existing text that tries to explain sub-unit overheads
