### 2018-07-30 b432cd2
Fix figure labels
