### 2018-07-30 ffb60c6
add documentation on running the unit tests
