### 2018-07-30 b9110c3
run the full tests, including doctests
