### 2018-07-30 c87c992
add docs for some of the exceptions
