### 2018-07-30 fd1e023
typo and wording adjustments
