+++ CONTRIBUTING.rst
@@ -132,12 +132,17 @@ Pull Request Guidelines
 
 Before you submit a pull request, check that it meets these guidelines:
 
-1. The pull request should include tests functionality that is not already
+1. The pull request should include tests for functionality that is not already
    tested. We strive for 100% test coverage and pull requests should not add any
-   new untested code.
-2. If the pull request adds functionality, the docs should be updated. Put
-   your new functionality into a function with a docstring, and add the
-   feature to the list in README.rst.
+   new untested code. Use the codecov.io reports on the pull request to gaugue
+   coverage. You can also generate coverage reports locally by running the
+   ``tox`` tests.
+2. If the pull request adds functionality the docs should be updated. If your
+   new functionality adds new functions or classes to the public API, please add
+   docstrings. If you modified an existing function or class in the public API,
+   please update the existing docstrings. If you modify private implementation
+   details, please use your judgment on documenting it with comments or
+   docstrings.
 3. The pull request should work for Python 2.7, 3.4, 3.5, 3.6, and 3.7. Check
    https://travis-ci.org/yt-project/unyt/pull_requests
    and make sure that the tests pass for all supported Python versions.
