### 2018-07-31 2c2c791
add docstrings for the rest of the exceptions
