+++ unyt/exceptions.py
@@ -24,10 +24,9 @@ class UnitOperationError(ValueError):
     >>> 3*u.g + 4*u.m  # doctest: +ELLIPSIS +NORMALIZE_WHITESPACE
     Traceback (most recent call last):
     ...
-    unyt.exceptions.UnitOperationError: The <ufunc 'add'> operator for
-    unyt_arrays with units "g" (dimensions "(mass)") and
+    unyt.exceptions.UnitOperationError: The <ufunc 'add'> operator
+    for unyt_arrays with units "g" (dimensions "(mass)") and
     "m" (dimensions "(length)") is not well defined.
-
     """
     def __init__(self, operation, unit1, unit2=None):
         self.operation = operation
@@ -57,8 +56,8 @@ class UnitConversionError(Exception):
     >>> data.to('m')  # doctest: +ELLIPSIS +NORMALIZE_WHITESPACE
     Traceback (most recent call last):
     ...
-    unyt.exceptions.UnitConversionError: Cannot convert between g (dim (mass))
-    and m (dim (length)).
+    unyt.exceptions.UnitConversionError: Cannot convert between 'g'
+    (dim '(mass)') and 'm' (dim '(length)').
     """
     def __init__(self, unit1, dimension1, unit2, dimension2):
         self.unit1 = unit1
@@ -68,8 +67,9 @@ class UnitConversionError(Exception):
         Exception.__init__(self)
 
     def __str__(self):
-        err = ("Cannot convert between %s (dim %s) and %s (dim %s)." %
-               (self.unit1, self.dimension1, self.unit2, self.dimension2))
+        err = ("Cannot convert between '%s' (dim '%s') and '%s' "
+               "(dim '%s')." % (self.unit1, self.dimension1, self.unit2,
+                                self.dimension2))
         return err
 
 
@@ -90,8 +90,8 @@ class MissingMKSCurrent(Exception):
     >>> us[ampere.dimensions]  # doctest: +ELLIPSIS +NORMALIZE_WHITESPACE
     Traceback (most recent call last):
     ...
-    unyt.exceptions.MissingMKSCurrent: The cgs unit system does not have a MKS
-    current base unit
+    unyt.exceptions.MissingMKSCurrent: The cgs unit system does not
+    have a MKS current base unit
 
     """
     def __init__(self, unit_system_name):
@@ -107,14 +107,14 @@ class MKSCGSConversionError(Exception):
     """Raised when conversion between MKS and CGS units cannot be performed
 
     This error is raised and caught internally and will expose itself
-    to the level of a user as part of a chained exception leading to a
+    to the user as part of a chained exception leading to a
     UnitConversionError.
     """
     def __init__(self, unit):
         self.unit = unit
 
     def __str__(self):
-        err = ("The %s unit cannot be safely converted." % self.unit)
+        err = ("The '%s' unit cannot be safely converted." % self.unit)
         return err
 
 
@@ -130,8 +130,8 @@ class UnitsNotReducible(Exception):
     Traceback (most recent call last):
     ...
     unyt.exceptions.UnitsNotReducible: The unit "A/cm" (dimensions
-    "(current_mks)/(length)") cannot be reduced to an expression within
-    the cgs system of units.
+    "(current_mks)/(length)") cannot be reduced to an expression
+    within the cgs system of units.
     """
     def __init__(self, unit, units_base):
         self.unit = unit
@@ -145,20 +145,21 @@ class UnitsNotReducible(Exception):
         return err
 
 
-class EquivalentDimsError(UnitOperationError):
-    def __init__(self, old_units, new_units, base):
[…]
