### 2018-08-03 679a447
update release notes
