### 2018-08-03 #809 @ngoldbaum
@kyleniemeyer I've pushed `unyt 1.0.5` to pypi and conda-forge. Can you take a look? If you approve I'll put that version on zenodo.
