### 2018-08-06 #809 @kyleniemeyer
@ngoldbaum OK, I have some edits for the paper:

- Please add more complete location information to the affiliations (e.g., city, state, country)
- First sentence in 2nd paragraph: capitalize "Python". Also, in the 7th paragraph.
- Please replace the hyphens used in the 2nd, 3rd, and 7th paragraphs with proper em-dashes (i.e., — vs -)
- "h5py" should probably also be in `texttt` format to match the rest of the software. Also, ndarray in the 4th paragraph and in the Pint section.
- I think there is a typo in this sentence of the 3rd paragraph: "In addition, unyt ships with a number of other useful predefined unit systems **based**, including ..."  (emphasis mine). The list structure in the second have of this sentence might need another look as well, as I think you have multiple lists that all use commas—if that is correct, you should use semicolons for the main list
- First sentence of the "NumPy ufunc performance" section: "Lastly, In Figures..." -> "Lastly, in Figures..."
- Reference Croton 2013: is this an article? It is missing the journal name (or equivalent)
- Reference Pontzen 2013: missing a URL or DOI
- The Astropy Collaboration reference is missing the arXiv identifier and link (I recommend citing in this format: https://arxiv.org/help/faq/references)

After you make these changes, please archive the entire repo to Zenodo and report back the DOI
