### 2018-08-13 f535da8
respond to kyle's comments, update codemeta
