+++ paper/paper.md
@@ -23,9 +23,9 @@ authors:
     orcid: 0000-0003-4423-0660
     affiliation: 2
 affiliations:
-  - name: National Center for Supercomputing Applications, University of Illinois at Urbana-Champaign
+  - name: National Center for Supercomputing Applications, University of Illinois at Urbana-Champaign. 1205 W Clark St, Urbana, IL USA 61801
     index: 1
-  - name: Harvard-Smithsonian Center for Astrophysics
+  - name: Harvard-Smithsonian Center for Astrophysics. 60 Garden St, Cambridge, MA USA 02138
     index: 2
 date: 24 May 2018
 bibliography: paper.bib
@@ -46,19 +46,19 @@ Mars due to the use of English Imperial units rather than metric units in the
 spacecraft control software [@nasa1999].
 
 The `unyt` library is designed both to aid quick calculations at an interactive
-python prompt and to be tightly integrated into a larger Python application or
+Python prompt and to be tightly integrated into a larger Python application or
 library. The top-level `unyt` namespace ships with a large number of predefined
 units and physical constants to aid setting up quick calculations without
 needing to look up unit data or the value of a physical constant. Using the
 `unyt` library as an interactive calculation aid only requires knowledge of
 basic Python syntax and awareness of a few of the methods of the `unyt_array`
-class - for example, the `unyt_array.to()` method to convert data to a different
+class — for example, the `unyt_array.to()` method to convert data to a different
 unit. As the complexity of the usage increases, `unyt` provides a number of
 optional features to aid these cases, including custom unit registries
 containing both predefined physical units as well as user-defined units,
-built-in output to disk via the pickle protocol and to HDF5 files using the h5py
-library [@h5py], and round-trip converters for unit objects defined by other
-popular Python unit libraries.
+built-in output to disk via the pickle protocol and to HDF5 files using the
+`h5py` library [@h5py], and round-trip converters for unit objects defined by
+other popular Python unit libraries.
 
 Physical units in the `unyt` class are defined in terms of the dimensions of the
 unit, a string representation, and a floating point scaling to the MKS unit
@@ -68,17 +68,17 @@ manipulation. The `unyt.Unit` object can represent arbitrary units formed out of
 base dimensions in the SI unit system: time, length, mass, temperature,
 luminance, and electric current. We currently treat units such as mol with the
 seventh SI base dimension, amount of substance, as dimensionless, although we
-are open to changing this based on feedback from users. In addition, `unyt`
-supports forming quantities defined in other unit systems - in particular CGS
+are open to changing this based on feedback from users. The `unyt` library
+supports forming quantities defined in other unit systems — in particular CGS
 Gaussian units common in astrophysics as well as geometrized "natural" units
 common in relativistic calculations. In addition, `unyt` ships with a number of
-other useful predefined unit systems based, including imperial units, Planck
-units, a unit system for calculations in the solar system, and a "galactic" unit
+other useful predefined unit systems including imperial units; Planck
+units; a unit system for calculations in the solar system; and a "galactic" unit
 system based on the solar mass, kiloparsecs, and Myr, a convention common in
 galactic astronomy.
 
 In addition to the `unyt.Unit` class, `unyt` also provides a two subclasses of
-the NumPy [@NumPy] ndarray [@vanderwalt2011], `unyt.unyt_array` and
+the NumPy [@NumPy] `ndarray` [@vanderwalt2011], `unyt.unyt_array` and
 `unyt.unyt_quantity` to represent arrays and scalars with units attached,
 respectively. The `unyt` library also provides a `unyt.UnitRegistry` class to
 allow custom systems of units, for example to track the internal unit system
[…]
