+++ paper/paper.bib
@@ -79,7 +79,8 @@ archivePrefix = "arXiv",
      R. and {Reed}, D.~M. and {Coles}, J. and {Quinn}, T.~R.},
   title = "{pynbody: Astrophysics Simulation Analysis for Python}",
   note = {Astrophysics Source Code Library, ascl:1305.002},
-  year = 2013
+  year = 2013,
+  howpublished = {\url{http://ascl.net/1305.002}}
 }
 @ARTICLE{astropy,
    author = {{The Astropy Collaboration} and {Price-Whelan}, A.~M. and {Sip{\H o}cz}, B.~M. and
@@ -130,7 +131,8 @@ archivePrefix = "arXiv",
      year = 2018,
     month = jan,
    adsurl = {http://adsabs.harvard.edu/abs/2018arXiv180102634T},
-  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+  adsnote = {Provided by the SAO/NASA Astrophysics Data System},
+  howpublished = {\url{https://arxiv.org/abs/1801.02634}}
 }
 @misc{Pint,
  author = {Grecco, Hernan E.},
@@ -215,7 +217,7 @@ month={May},}
 @ARTICLE{croton2013,
    author = {{Croton}, D.~J.},
     title = "{Damn You, Little h! (Or, Real-World Applications of the Hubble Constant Using Observed and Simulated Data)}",
-  journal = {\pasa},
+  journal = {Publications of the Astronomical Society of Australia},
 archivePrefix = "arXiv",
    eprint = {1308.4150},
  keywords = {cosmology, galaxies, methods: observational},
