### 2018-08-13 #809 @augustfly
Hi,

I'm not sure what is being asked of me?

Looking at the last version of the bibtex before the last commit, and at the JOSS author guidelines, the only problem I see is that `pynbody` wasn't a recent ADS bibtex entry.

The guidelines --

https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography

suggest ADS macros (e.g., `\pasa`) are supported as are `eprint` + `archiveprefix` keys, which is how to encode both arXiv and ASCL entries into anchored URLs (as described here: https://arxiv.org/hypertex/bibstyles/).

FWIW, `howpublished` is a bibtex hack that I hope isn't part of anyone's workflow. :-)  Not that I'm not one to avoid bibtex hacks, having added `version` tags to our Journal's preferred bibtex styles for citing software.

having said all that -- the person you want is @aaccomazzi
