### 2018-08-13 #809 @aaccomazzi
Just checked the generated PDF and saw no problems with missing links in the reference, so I assume the issue has been fixed by the authors.  If there is a systematic issue with ADS reference generation please file an issue with full details here: https://github.com/adsabs/export_service
